requirements.txt
blinker==1.3
chardet==2.3.0
colorama==0.3.2
Flask==0.10.1
gpiozero==1.3.2
html5lib==0.999
itsdangerous==0.24
Jinja2==2.7.3
luma.core==0.7.4
luma.oled==2.2.7
lxkeymap==0.1
MarkupSafe==0.23
mcpi==0.1.1
microdotphat==0.1.3
monotonic==1.3
ndg-httpsclient==0.3.2
numpy==1.8.2
olefile==0.44
picamera==1.13
picraft==1.0
pifacecommon==4.2.1
pifacedigitalio==3.1.0
pigpio==1.35
Pillow==4.1.0
pyasn1==0.1.7
pygame==1.9.2a0
pygobject==3.14.0
pyinotify==0.9.4
pyOpenSSL==0.13.1
pyserial==2.6
python-apt==0.9.3.12
requests==2.4.3
rpi-ws281x==1.0.0
RPi.GPIO==0.6.3
RTIMULib==7.2.1
sense-emu==1.0
sense-hat==2.2.0
six==1.8.0
smbus==1.1
smbus2==0.1.4
spidev==3.2
twython==3.1.2
urllib3==1.9.1
Werkzeug==0.9.6
